Logged user for CRUD contacts

+++ controllers/contactController.js
@@ -3,16 +3,17 @@ const Contact = require("../models/contactModel");
 
 //! Get all contacts
 const getAllContact = asyncHandler(async (req, res) => {
-  res.status(200).json({
-    message: "Get all contacts from the backend!",
+  const contact = await Contact.find({
+    user_id: req.user.id,
   });
+  res.status(200).json(contact);
 });
 
 //! Get single contact
 const getContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
 
-  if(!contact) {
+  if (!contact) {
     res.status(404);
     throw new Error("Contact not found");
   }
@@ -33,6 +34,7 @@ const createContact = asyncHandler(async (req, res) => {
     name,
     email,
     phone,
+    user_id: req.user.id,
   });
 
   res.status(200).json(contact);
@@ -42,9 +44,16 @@ const createContact = asyncHandler(async (req, res) => {
 const updateContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
 
-  if(!contact) {
+  if (!contact) {
     res.status(404);
     throw new Error("Contact not found");
+  }
+
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error(
+      "You are not authorized to update this contact information"
+    );
   }
 
   const updateContact = await Contact.findByIdAndUpdate(
@@ -54,7 +63,7 @@ const updateContact = asyncHandler(async (req, res) => {
       new: true,
       runValidators: true,
     }
-  )
+  );
 
   res.status(200).json(updateContact);
 });
@@ -63,9 +72,14 @@ const updateContact = asyncHandler(async (req, res) => {
 const deleteContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
 
-  if(!contact) {
+  if (!contact) {
     res.status(404);
     throw new Error("Contact not found");
+  }
+
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("You are not authorized to delete this contact");
   }
 
   const deleteContact = await Contact.findByIdAndDelete(req.params.id);
